Moved provider related functions to ProviderManagementIntrinsics
